requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
PyPDF2
scikit-learn
nltk
tiktoken
openai
uuid
